fmt and clippy

+++ src/util/id/meid.rs
@@ -37,7 +37,7 @@ pub fn gen_meid(time: u64) -> String {
 
 pub fn parse_meid(id: &str) -> Result<SystemTime, ParseIntError> {
     let timestamp = u64::from_str_radix(&id[0..12], 16).unwrap() - 0x800000000000;
-    let time = SystemTime::UNIX_EPOCH + std::time::Duration::from_millis(timestamp as u64);
+    let time = SystemTime::UNIX_EPOCH + std::time::Duration::from_millis(timestamp);
     Ok(time)
 }
